+++ packages/edron/src/edron/tooling.py
@@ -237,8 +237,8 @@ def doctor(
 
     requirements = {
         "edron": (">=0.9,<0.10", "edron"),
-        "hedron": (">=0.67.0,<0.68", "hedron"),
-        "hedron-data": (">=0.67.0,<0.68", "hedron_data"),
+        "hedron": (">=0.67.0,<2.0", "hedron"),
+        "hedron-data": (">=0.67.0,<2.0", "hedron_data"),
         "hedron-charts": (">=0.2,<0.3", "hedron_charts"),
         "hedron-maps": (">=0.1,<0.2", "hedron_maps"),
     }